refactor: created index files for each component
