+++ src/assets/index.js
@@ -1,0 +1,21 @@
+// icons
+export { default as FacebookIcon } from "./icons/FacebookIcon.svg";
+export { default as icon1 } from "./icons/icon1.svg";
+export { default as icon2 } from "./icons/icon2.svg";
+export { default as icon3 } from "./icons/icon3.svg";
+export { default as LinkedinIcon } from "./icons/LinkedinIcon.svg";
+export { default as logoIcon } from "./icons/logo.svg";
+export { default as programmerIcon } from "./icons/programmer.svg";
+export { default as reactIcon } from "./icons/react.svg";
+export { default as toggleButtonIcon } from "./icons/toggleButton.svg";
+export { default as TwitterIcon } from "./icons/TwitterIcon.svg";
+export { default as WhatsappIcon } from "./icons/WhatsappIcon.svg";
+
+// images
+export { default as backEnd } from "./images/back-end.jpg";
+export { default as blurImg } from "./images/blur.jpg";
+export { default as frontendImg } from "./images/Front-end.jpg";
+export { default as imgRutes } from "./images/img_rutes.jpg";
+export { default as notFoundImg } from "./images/NotFound.png";
+export { default as webImg } from "./images/Web.png";
+export { default as youtubeImg } from "./images/youTube.png";
