+++ src/components/AboutUs/AboutUs.css
@@ -1,5 +1,5 @@
 .AboutUs {
-  background-image: url("../../assets/background-gray.png");
+  background-image: url("/background-gray.png");
   background-repeat: no-repeat;
   background-size: 100%;
   width: 100%;
