+++ src/components/AboutUs/index.js
@@ -1,0 +1,1 @@
+export { default } from './AboutUs'
