+++ src/components/CardTopics/CardTopics.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Link } from "react-router-dom";
 
 const CardTopics = ({ title, description, link }) => {
