+++ src/components/MainCover/MainCover.jsx
@@ -1,5 +1,6 @@
 import React from "react";
-import imageCoverProgrammer from "../../assets/icons/programmer.svg";
+import { programmerIcon as imageCoverProgrammer } from "../../assets";
+
 import "./MainCover.css";
 
 const MainCover = () => {
