+++ src/components/Map/Map.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import "./Map.css";
 
 const Map = () => {
