+++ src/components/Modal/Modal.jsx
@@ -1,9 +1,9 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import { createPortal } from "react-dom";
+
 import "./Modal.css";
 
 const Modal = ({ children }) => {
-  return ReactDOM.createPortal(
+  return createPortal(
     <div className="ModalBackground">{children}</div>,
     document.getElementById("modal")
   );
