+++ src/components/Topics/Topics.jsx
@@ -1,5 +1,3 @@
-import React from "react";
-import { Link } from "react-router-dom";
 import "./Topics.css";
 
 const Topics = ({ title, ide, children }) => {
