+++ src/components/Utils/Utils.jsx
@@ -1,5 +1,5 @@
-import React from "react";
 import { Link, useParams } from "react-router-dom";
+
 import "./Utils.css";
 
 export const Title = ({ title, styles = "" }) => {
@@ -27,5 +27,5 @@ export const ItemContent = ({ item }) => {
 };
 
 export const Loader = () => {
-  return <span class="loader"></span>;
+  return <span className="loader"></span>;
 };
