+++ src/layouts/FormLayout/FormLayout.jsx
@@ -1,5 +1,5 @@
-import React from "react";
 import { Outlet } from "react-router-dom";
+
 import "./FormLayout.css";
 
 const FormLayout = () => {
