+++ src/layouts/MainLayout/MainLayout.jsx
@@ -1,8 +1,8 @@
-import React from "react";
 import { Outlet } from "react-router-dom";
-import Header from "../../components/Header/Header";
-import Footer from "../../components/Footer/Footer";
-import FloatingButtons from "../../components/FloatingButtons/FloatingButtons";
+import Header from "../../components/Header";
+import Footer from "../../components/Footer";
+import FloatingButtons from "../../components/FloatingButtons";
+
 import "./MainLayout.css";
 
 const MainLayout = () => {
