+++ src/pages/Home/Home.jsx
@@ -1,9 +1,10 @@
-import Complement from "../../components/Complement/Complement";
-import Map from "../../components/Map/Map";
-import MainCover from "../../components/MainCover/MainCover";
-import Information from "../../components/Information/Information";
-import AboutUs from "../../components/AboutUs/AboutUs";
-import Service from "../../components/Service/Service";
+import Complement from "../../components/Complement";
+import Map from "../../components/Map";
+import MainCover from "../../components/MainCover";
+import Information from "../../components/Information";
+import AboutUs from "../../components/AboutUs";
+import Service from "../../components/Service";
+
 import "./Home.css";
 
 const Home = () => {
