+++ vite.config.js
@@ -4,6 +4,10 @@ import react from '@vitejs/plugin-react'
 export default defineConfig({
   plugins: [react()],
   server: {
-    host: true
-  }
+    host: true,
+    port: 3000,
+  },
+  preview: {
+    port: 3200
+  },
 })
